Update minimal requirement of six to >= 1.12.0
urllib3-1.25.7 do not work with python3-six-1.11.0 which
is in CentOS8 base repo. So to ensure when urllib3 is installed
already installed python3-six is upgraded let's set minimum
requirement of six to >= 1.12.0[1] which contains required methods like
six.ensure_text method needed by urllib3-1.25.7.

PR proposed to update in Fedora too[2].

In RDO we have six==1.14.0 in CentOS8 repos currently, so we
are good.

[1] https://github.com/urllib3/urllib3/commit/5b047b
[2] https://src.fedoraproject.org/rpms/python-urllib3/pull-request/10